requirements.txt
Django==1.11.29
image==1.5.33
Pillow==6.2.2
pytz==2021.1
qrcode==6.1
six==1.16.0
xeger==0.3.5
